requirements.txt
django-cors-headers==3.10.0
requests==2.26.0
json5==0.9.6
jsonschema==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
